Moving IRegressionAlgorithmDefinition to QuantConnect.Common
