+++ Algorithm.CSharp/AddRemoveOptionUniverseRegressionAlgorithm.cs
@@ -20,6 +20,7 @@ using System.Linq;
 using QuantConnect.Data;
 using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Util;
+using QuantConnect.Interfaces;
 
 // ReSharper disable InvokeAsExtensionMethod -- .net 4.7.2 added ToHashSet and it looks like our version of mono has it as well causing ambiguity in the cloud
