+++ Algorithm.CSharp/BasicTemplateAlgorithm.cs
@@ -15,6 +15,7 @@
 
 using System.Collections.Generic;
 using QuantConnect.Data;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
