+++ Algorithm.CSharp/CoarseFundamentalTop5Algorithm.cs
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-using System;
+using QuantConnect.Interfaces;
 using System.Collections.Generic;
 using System.Linq;
 using QuantConnect.Data.Market;
