+++ Algorithm.CSharp/ConvertToFrameworkAlgorithm.cs
@@ -19,6 +19,7 @@ using QuantConnect.Algorithm.Framework;
 using QuantConnect.Algorithm.Framework.Alphas;
 using QuantConnect.Data.Market;
 using QuantConnect.Indicators;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
