+++ Algorithm.CSharp/CustomDataRegressionAlgorithm.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Collections.Generic;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
