+++ Algorithm.CSharp/HourSplitRegressionAlgorithm.cs
@@ -15,6 +15,7 @@
 
 using System.Collections.Generic;
 using QuantConnect.Data.Market;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
