+++ Algorithm.CSharp/IndicatorSuiteAlgorithm.cs
@@ -19,6 +19,7 @@ using QuantConnect.Data;
 using QuantConnect.Data.Custom;
 using QuantConnect.Data.Market;
 using QuantConnect.Indicators;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
