+++ Algorithm.CSharp/MACDTrendAlgorithm.cs
@@ -17,6 +17,7 @@ using System;
 using System.Collections.Generic;
 using QuantConnect.Data.Market;
 using QuantConnect.Indicators;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
