+++ Algorithm.CSharp/OptionOpenInterestRegressionAlgorithm.cs
@@ -18,6 +18,7 @@ using System;
 using System.Collections.Generic;
 using QuantConnect.Data;
 using QuantConnect.Orders;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
