+++ Algorithm.CSharp/QuantConnect.Algorithm.CSharp.csproj
@@ -94,7 +94,6 @@
     <Compile Include="ConvertToFrameworkAlgorithm.cs" />
     <Compile Include="RawPricesCoarseUniverseAlgorithm.cs" />
     <Compile Include="CompositeAlphaModelFrameworkAlgorithm.cs" />
-    <Compile Include="IRegressionAlgorithmDefinition.cs" />
     <Compile Include="SectorExposureRiskFrameworkAlgorithm.cs" />
     <Compile Include="DuplicateSecurityWithBenchmarkRegressionAlgorithm.cs" />
     <Compile Include="BasicTemplateCryptoAlgorithm.cs" />
