+++ Algorithm.CSharp/UpdateOrderRegressionAlgorithm.cs
@@ -21,6 +21,7 @@ using QuantConnect.Data;
 using QuantConnect.Orders;
 using QuantConnect.Securities;
 using QuantConnect.Util;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
