+++ Algorithm.CSharp/WeeklyUniverseSelectionRegressionAlgorithm.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.Linq;
 using QuantConnect.Data.Market;
 using QuantConnect.Data.UniverseSelection;
+using QuantConnect.Interfaces;
 
 namespace QuantConnect.Algorithm.CSharp
 {
